Add comprehensive tests for move picker functionality
- Implement basic move picker tests covering stages, quiescence, PV moves, and killer moves.
- Introduce capture-related tests to validate SEE calculations and MVV-LVA ordering.
- Create drop-related tests to enforce pawn drop restrictions and validate legality.
- Add tests for uchifuzume scenarios, ensuring correct handling of checkmate by pawn drop.
- Define types for scored moves and move picker stages to enhance move generation efficiency.
