--- packages/rust-core/crates/engine-core/src/movegen/move_picker.rs
@@ -1,1521 +1,0 @@
-//! Move picker for staged move generation and ordering
-//!
-//! Generates moves in stages for better search efficiency:
-//! 1. TT move
-//! 2. Good captures (MVV-LVA)
-//! 3. Killer moves
-//! 4. Bad captures
-//! 5. Quiet moves (history ordered)
-
-use crate::search::types::SearchStack;
-use crate::shogi::{attacks, Move, MoveList};
-use crate::{Bitboard, Color, History, MoveGen, PieceType, Position, Square};
-
-/// Scored move for ordering
-#[derive(Clone, Copy, Debug)]
-struct ScoredMove {
-    mv: Move,
-    score: i32,
-}
-
-impl ScoredMove {
-    fn new(mv: Move, score: i32) -> Self {
-        ScoredMove { mv, score }
-    }
-}
-
-/// Stage of move generation
-///
-/// The order is carefully chosen based on the probability of causing a beta cutoff:
-/// 1. TT move - Most likely to be the best move from previous search
-/// 2. Good captures - Positive SEE captures are often good moves
-/// 3. Killers - Moves that caused cutoffs at the same depth
-/// 4. Quiet moves - Non-captures ordered by history heuristic
-/// 5. Bad captures - Negative SEE captures (least likely to be good)
-///
-/// Bad captures are intentionally placed last because:
-/// - They have negative SEE value (lose material)
-/// - They rarely cause beta cutoffs
-/// - In quiescence search, they are often skipped entirely
-/// - Late Move Reductions (LMR) work better with bad moves at the end
-///
-/// This ordering matches strong engines like Stockfish and maximizes
-/// the efficiency of alpha-beta pruning.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum MovePickerStage {
-    /// Root PV move (highest priority at root node)
-    RootPV,
-    /// TT move
-    TTMove,
-    /// Generate captures
-    GenerateCaptures,
-    /// Good captures (SEE >= 0)
-    GoodCaptures,
-    /// Killer moves
-    Killers,
-    /// Generate quiet moves
-    GenerateQuiets,
-    /// All quiet moves
-    QuietMoves,
-    /// Bad captures (SEE < 0)
-    BadCaptures,
-    /// End of moves
-    End,
-}
-
-/// Move picker for efficient move ordering
-pub struct MovePicker<'a> {
-    /// Current position
-    pos: Position,
-    /// TT move
-    tt_move: Option<Move>,
-    /// PV move (from previous iteration)
-    pv_move: Option<Move>,
-    /// History heuristics
-    history: &'a History,
-    /// Search stack entry
-    stack: SearchStack,
-    /// Current stage
-    stage: MovePickerStage,
-    /// Generated moves
-    moves: Vec<ScoredMove>,
-    /// Bad captures
-    bad_captures: Vec<ScoredMove>,
-    /// Current index in moves/killers/bad_captures depending on stage
-    /// - In Killers stage: index into stack.killers[] (0-1)
-    /// - In BadCaptures stage: index into bad_captures vector
-    /// - Reset to 0 when transitioning to a new stage
-    current: usize,
-    /// Skip quiet moves (for quiescence search)
-    skip_quiets: bool,
-}
-
-impl<'a> MovePicker<'a> {
-    /// Create new move picker for main search
-    pub fn new(
-        pos: &Position,
-        tt_move: Option<Move>,
-        pv_move: Option<Move>,
-        history: &'a History,
-        stack: &SearchStack,
-        ply: usize,
-    ) -> Self {
-        // Validate PV move: check for duplicates and legality
-        let validated_pv_move = pv_move.filter(|&mv| {
-            // Skip if same as TT move
-            tt_move != Some(mv) &&
-            // Check legality
-            pos.is_legal_move(mv)
-        });
-
-        // Determine starting stage based on root node and PV availability
-        let stage = if ply == 0 && validated_pv_move.is_some() {
-            MovePickerStage::RootPV
-        } else {
-            MovePickerStage::TTMove
-        };
-
-        MovePicker {
-            pos: pos.clone(),
-            tt_move,
-            pv_move: validated_pv_move,
-            history,
-            stack: stack.clone(),
-            stage,
-            moves: Vec::new(),
-            bad_captures: Vec::new(),
-            current: 0, // Initialize to 0, will be used differently in each stage
-            skip_quiets: false,
-        }
-    }
-
-    /// Create new move picker for quiescence search (captures only)
-    ///
-    /// In quiescence search, we skip quiet moves and bad captures.
-    /// Only good captures (SEE >= 0) are considered to avoid search explosion.
-    /// This is why bad captures are placed after quiet moves in normal search -
-    /// they're often not searched at all.
-    pub fn new_quiescence(
-        pos: &Position,
-        tt_move: Option<Move>,
-        history: &'a History,
-        stack: &SearchStack,
-        ply: usize,
-    ) -> Self {
-        let mut picker = Self::new(pos, tt_move, None, history, stack, ply);
-        picker.skip_quiets = true;
-        picker
-    }
-
-    /// Get next move
-    pub fn next_move(&mut self) -> Option<Move> {
-        loop {
-            match self.stage {
-                MovePickerStage::RootPV => {
-                    self.stage = MovePickerStage::TTMove;
-                    if let Some(pv_move) = self.pv_move {
-                        return Some(pv_move);
-                    }
-                }
-
-                MovePickerStage::TTMove => {
-                    self.stage = MovePickerStage::GenerateCaptures;
-                    if let Some(tt_move) = self.tt_move {
-                        if self.pos.is_legal_move(tt_move) {
-                            return Some(tt_move);
-                        }
-                    }
-                }
-
-                MovePickerStage::GenerateCaptures => {
-                    self.generate_captures();
-                    self.score_captures();
-                    self.stage = MovePickerStage::GoodCaptures;
-                }
-
-                MovePickerStage::GoodCaptures => {
-                    if let Some(mv) = self.pick_best() {
-                        if Some(mv) != self.tt_move && Some(mv) != self.pv_move {
-                            // Separate good and bad captures
-                            let score = self.see(mv);
-                            if score >= 0 {
-                                return Some(mv);
-                            } else {
-                                // Save bad capture for later
-                                self.bad_captures.push(ScoredMove::new(mv, score));
-                            }
-                        }
-                    } else {
-                        self.stage = if self.skip_quiets {
-                            MovePickerStage::End
-                        } else {
-                            MovePickerStage::Killers
-                        };
-                        self.current = 0; // Reset index for killer moves iteration
-                    }
-                }
-
-                MovePickerStage::Killers => {
-                    if self.current < 2 {
-                        // Check up to 2 killer moves
-                        if let Some(killer) = self.stack.killers[self.current] {
-                            self.current += 1; // Move to next killer slot
-                            if Some(killer) != self.tt_move
-                                && Some(killer) != self.pv_move
-                                && !self.is_capture(killer)
-                                && self.pos.is_legal_move(killer)
-                            {
-                                return Some(killer);
-                            }
-                        } else {
-                            self.current += 1; // Skip empty killer slot
-                        }
-                    } else {
-                        // Transition from Killers to GenerateQuiets
-                        // Next we'll generate and score quiet moves
-                        self.stage = MovePickerStage::GenerateQuiets;
-                        self.current = 0; // Not used in GenerateQuiets, but good practice to reset
-                    }
-                }
-
-                MovePickerStage::GenerateQuiets => {
-                    self.generate_quiets();
-                    self.score_quiets();
-                    // Transition to QuietMoves (not BadCaptures)
-                    // This is intentional: quiet moves with good history scores
-                    // are more likely to be good than losing captures
-                    self.stage = MovePickerStage::QuietMoves;
-                    self.current = 0; // Not used in QuietMoves (pick_best manages its own index)
-                }
-
-                MovePickerStage::QuietMoves => {
-                    if let Some(mv) = self.pick_best() {
-                        // No need to check for TT move or killers - already filtered during generation
-                        return Some(mv);
-                    } else {
-                        // Bad captures come last - they rarely produce good moves
-                        // and are often pruned by Late Move Reductions
-                        self.stage = MovePickerStage::BadCaptures;
-                        self.current = 0; // Reset index to iterate through bad_captures vector
-                    }
-                }
-
-                MovePickerStage::BadCaptures => {
-                    if self.current < self.bad_captures.len() {
-                        let mv = self.bad_captures[self.current].mv;
-                        self.current += 1; // Move to next bad capture
-                        if Some(mv) != self.tt_move && Some(mv) != self.pv_move {
-                            return Some(mv);
-                        }
-                    } else {
-                        self.stage = MovePickerStage::End;
-                    }
-                }
-
-                MovePickerStage::End => {
-                    return None;
-                }
-            }
-        }
-    }
-
-    /// Generate capture moves
-    fn generate_captures(&mut self) {
-        self.moves.clear();
-        let mut move_list = MoveList::new();
-        let mut gen = MoveGen::new();
-        gen.generate_captures(&self.pos, &mut move_list);
-
-        for &mv in move_list.as_slice() {
-            self.moves.push(ScoredMove::new(mv, 0));
-        }
-    }
-
-    /// Generate quiet moves
-    fn generate_quiets(&mut self) {
-        self.moves.clear();
-        let mut move_list = MoveList::new();
-        let mut gen = MoveGen::new();
-        gen.generate_all(&self.pos, &mut move_list);
-
-        // Add only non-captures that are not killers, TT move, or PV move
-        for &mv in move_list.as_slice() {
-            if !self.is_capture(mv)
-                && Some(mv) != self.tt_move
-                && Some(mv) != self.pv_move
-                && !self.is_killer(mv)
-            {
-                self.moves.push(ScoredMove::new(mv, 0));
-            }
-        }
-    }
-
-    /// Score captures using Static Exchange Evaluation (SEE)
-    fn score_captures(&mut self) {
-        for i in 0..self.moves.len() {
-            let mv = self.moves[i].mv;
-            if self.get_captured_piece(mv).is_some() {
-                // Calculate SEE value for this capture
-                let see_value = self.see(mv);
-
-                // Use SEE value as the primary score
-                self.moves[i].score = see_value;
-
-                // Small promotion bonus (SEE already accounts for promoted piece value)
-                if mv.is_promote() {
-                    self.moves[i].score += 100;
-                }
-            }
-        }
-    }
-
-    /// Score quiet moves using history
-    fn score_quiets(&mut self) {
-        for scored_move in &mut self.moves {
-            let mv = scored_move.mv;
-            scored_move.score = self.history.get_score(self.pos.side_to_move, mv, None);
-
-            // Promotion bonus
-            if mv.is_promote() {
-                scored_move.score += 300;
-            }
-        }
-    }
-
-    /// Pick best move from current list
-    fn pick_best(&mut self) -> Option<Move> {
-        if self.current >= self.moves.len() {
-            return None;
-        }
-
-        // Find best remaining move
-        let best_idx = self.current
-            + self.moves[self.current..]
-                .iter()
-                .enumerate()
-                .max_by_key(|(_, m)| m.score)
-                .map(|(i, _)| i)?;
-
-        // Swap with current position
-        self.moves.swap(self.current, best_idx);
-        let result = self.moves[self.current].mv;
-        self.current += 1;
-
-        Some(result)
-    }
-
-    /// Check if move is a capture
-    fn is_capture(&self, mv: Move) -> bool {
-        !mv.is_drop() && self.pos.board.piece_on(mv.to()).is_some()
-    }
-
-    /// Check if move is a killer
-    fn is_killer(&self, mv: Move) -> bool {
-        self.stack.killers[0] == Some(mv) || self.stack.killers[1] == Some(mv)
-    }
-
-    /// Get captured piece
-    fn get_captured_piece(&self, mv: Move) -> Option<PieceType> {
-        if mv.is_drop() {
-            None
-        } else {
-            self.pos.board.piece_on(mv.to()).map(|p| p.piece_type)
-        }
-    }
-
-    /// Static Exchange Evaluation
-    fn see(&self, mv: Move) -> i32 {
-        // Use the full SEE implementation from Position
-        self.pos.see(mv)
-    }
-}
-
-// Extension for Position to check legal moves
-impl Position {
-    /// Check if the current player has a pawn in the given file
-    fn has_pawn_in_file(&self, file: u8) -> bool {
-        self.has_pawn_in_file_for_color(file, self.side_to_move)
-    }
-
-    /// Check if the specified color has a pawn in the given file
-    fn has_pawn_in_file_for_color(&self, file: u8, color: Color) -> bool {
-        let pawn_bb = self.board.piece_bb[color as usize][PieceType::Pawn as usize];
-        let file_mask = attacks::file_mask(file);
-
-        // Get unpromoted pawns in this file
-        let unpromoted_pawns_in_file = pawn_bb & file_mask & !self.board.promoted_bb;
-
-        !unpromoted_pawns_in_file.is_empty()
-    }
-
-    /// Check if dropping a pawn would result in checkmate
-    fn is_checkmate_by_pawn_drop(&self, pawn_drop: Move) -> bool {
-        // The pawn must give check to the opponent's king
-        let defense_color = self.side_to_move.opposite();
-        let king_sq = match self.board.king_square(defense_color) {
-            Some(sq) => sq,
-            None => return false, // No king
-        };
-
-        // Check if pawn would give check (pawn attacks one square forward)
-        let pawn_sq = pawn_drop.to();
-
-        // Debug assert: pawn drops should already be filtered to valid positions
-        #[cfg(debug_assertions)]
-        {
-            debug_assert!(
-                !(self.side_to_move == Color::Black && pawn_sq.rank() == 0),
-                "Black pawn cannot be dropped on rank 1"
-            );
-            debug_assert!(
-                !(self.side_to_move == Color::White && pawn_sq.rank() == 8),
-                "White pawn cannot be dropped on rank 9"
-            );
-        }
-
-        // Black pawn attacks upward (toward rank 0), white pawn attacks downward (toward rank 8)
-        let expected_king_sq = if self.side_to_move == Color::Black {
-            // Black pawn at rank N attacks rank N-1
-            let rank = pawn_sq.rank();
-            if rank == 0 {
-                return false; // Black pawn at rank 0 cannot attack (invalid position)
-            }
-            Square::new(pawn_sq.file(), rank - 1)
-        } else {
-            // White pawn at rank N attacks rank N+1
-            let rank = pawn_sq.rank();
-            if rank == 8 {
-                return false; // White pawn at rank 8 cannot attack (invalid position)
-            }
-            Square::new(pawn_sq.file(), rank + 1)
-        };
-
-        if king_sq != expected_king_sq {
-            return false; // Pawn doesn't give check
-        }
-
-        // Step 1: Simulate pawn drop to check support and captures in the actual position
-        let mut test_pos = self.clone();
-        test_pos.do_move(pawn_drop);
-
-        // Check if the pawn has support (can't be captured by king)
-        if !test_pos.is_attacked(pawn_sq, self.side_to_move) {
-            return false; // The pawn has no followers, king can capture it
-        }
-
-        // Step 2: Check if opponent's pieces (except king/lance/pawn) can capture the pawn
-        let capture_candidates = test_pos.get_attackers_to(pawn_sq, defense_color);
-
-        // Exclude king, lance, and pawn from capture candidates
-        let king_bb = test_pos.board.piece_bb[defense_color as usize][PieceType::King as usize];
-        let lance_bb = test_pos.board.piece_bb[defense_color as usize][PieceType::Lance as usize];
-        let pawn_bb = test_pos.board.piece_bb[defense_color as usize][PieceType::Pawn as usize];
-        let excluded = king_bb | lance_bb | pawn_bb;
-
-        let valid_capture_candidates = capture_candidates & !excluded;
-
-        // Check for pinned pieces
-        let pinned = test_pos.get_blockers_for_king(defense_color);
-        let pawn_file_mask = Bitboard::file_mask(pawn_sq.file());
-        let not_pinned_for_capture = !pinned | pawn_file_mask;
-
-        let can_capture = valid_capture_candidates & not_pinned_for_capture;
-        if !can_capture.is_empty() {
-            return false; // Some piece can capture the pawn
-        }
-
-        // Step 3: Check if king can escape
-        let king_moves = attacks::king_attacks(king_sq);
-
-        // King cannot capture its own pieces (Shogi rule)
-        let friend_blocks = test_pos.board.occupied_bb[defense_color as usize];
-        let mut king_escape_candidates = king_moves & !friend_blocks;
-
-        // Remove the pawn square (king can't capture it due to support)
-        let mut pawn_sq_bb = Bitboard::EMPTY;
-        pawn_sq_bb.set(pawn_sq);
-        king_escape_candidates &= !pawn_sq_bb;
-
-        // Check each escape square
-        let mut candidates = king_escape_candidates;
-        while let Some(escape_sq) = candidates.pop_lsb() {
-            // Simulate king moving to escape square
-            let king_move = Move::normal(king_sq, escape_sq, false);
-            let mut escape_test_pos = test_pos.clone();
-
-            // Try to make the king move
-            // If there's a piece to capture, it will be handled by do_move
-            escape_test_pos.do_move(king_move);
-
-            // Check if king is safe after moving
-            let is_safe = !escape_test_pos.is_check(defense_color);
-
-            if is_safe {
-                return false; // King has a safe escape
-            }
-        }
-
-        // All conditions met - it's checkmate by pawn drop
-        true
-    }
-
-    /// Check if a move is legal
-    ///
-    /// This optimized version uses do_move/undo_move to check legality.
-    /// It's much faster than generating all legal moves (O(1) vs O(N)).
-    pub fn is_legal_move(&self, mv: Move) -> bool {
-        // Basic validation
-        if mv.is_drop() {
-            // Check if we have the piece to drop
-            let piece_type = mv.drop_piece_type();
-            let color_idx = self.side_to_move as usize;
-            let Some(hand_idx) = piece_type.hand_index() else {
-                return false; // Can't drop King or invalid type
-            };
-
-            if self.hands[color_idx][hand_idx] == 0 {
-                return false;
-            }
-
-            // Check if target square is empty
-            if self.board.piece_on(mv.to()).is_some() {
-                return false;
-            }
-
-            // Check piece-specific drop restrictions
-            match piece_type {
-                PieceType::Pawn => {
-                    let to_rank = mv.to().rank();
-
-                    // Check rank restrictions - pawn cannot be dropped on the last rank
-                    if (self.side_to_move == Color::Black && to_rank == 0)
-                        || (self.side_to_move == Color::White && to_rank == 8)
-                    {
-                        return false;
-                    }
-
-                    // Check nifu (double pawn)
-                    if self.has_pawn_in_file(mv.to().file()) {
-                        return false;
-                    }
-
-                    // Check uchifuzume (checkmate by pawn drop)
-                    if self.is_checkmate_by_pawn_drop(mv) {
-                        return false;
-                    }
-                }
-                PieceType::Lance => {
-                    let to_rank = mv.to().rank();
-
-                    // Check rank restrictions - lance cannot be dropped on the last rank
-                    if (self.side_to_move == Color::Black && to_rank == 0)
-                        || (self.side_to_move == Color::White && to_rank == 8)
-                    {
-                        return false;
-                    }
-                }
-                PieceType::Knight => {
-                    let to_rank = mv.to().rank();
-
-                    // Check rank restrictions - knight cannot be dropped on the last two ranks
-                    if (self.side_to_move == Color::Black && to_rank <= 1)
-                        || (self.side_to_move == Color::White && to_rank >= 7)
-                    {
-                        return false;
-                    }
-                }
-                _ => {} // Other pieces have no special drop restrictions
-            }
-        } else {
-            // Normal move
-            if let Some(from) = mv.from() {
-                // Check if there's a piece at the from square
-                if let Some(piece) = self.board.piece_on(from) {
-                    // Check if it's our piece
-                    if piece.color != self.side_to_move {
-                        return false;
-                    }
-
-                    // Check if capturing our own piece
-                    if let Some(to_piece) = self.board.piece_on(mv.to()) {
-                        if to_piece.color == self.side_to_move {
-                            return false;
-                        }
-                    }
-                } else {
-                    return false;
-                }
-            } else {
-                return false;
-            }
-        }
-
-        // Try to make the move and check if it leaves king in check
-        let mut test_pos = self.clone();
-        test_pos.do_move(mv);
-
-        // Check if the side that made the move left their king in check
-        let king_in_check = test_pos.is_check(self.side_to_move);
-
-        !king_in_check
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{usi::parse_usi_square, Board, Piece};
-
-    use super::*;
-
-    #[test]
-    fn test_move_picker_stages() {
-        let pos = Position::startpos();
-        let history = History::new();
-        let stack = SearchStack::default();
-
-        // Use a known legal move from starting position
-        // Black pawn at rank 6 moves toward rank 0
-        let tt_move = Some(Move::normal(
-            parse_usi_square("2g").unwrap(), // Black pawn
-            parse_usi_square("2f").unwrap(), // One square forward
-            false,
-        ));
-
-        let mut picker = MovePicker::new(&pos, tt_move, None, &history, &stack, 1);
-
-        // First move should be TT move
-        let first_move = picker.next_move();
-        assert_eq!(first_move, tt_move);
-
-        // Subsequent moves should not include TT move
-        let mut moves = Vec::new();
-        while let Some(mv) = picker.next_move() {
-            assert_ne!(Some(mv), tt_move);
-            moves.push(mv);
-        }
-
-        // Should have generated all legal moves except TT move
-        assert!(!moves.is_empty());
-    }
-
-    #[test]
-    fn test_quiescence_picker() {
-        // Create a position with no possible captures
-        let mut pos = Position::empty();
-
-        // Add kings (required)
-        pos.board
-            .put_piece(parse_usi_square("5i").unwrap(), Piece::new(PieceType::King, Color::Black));
-        pos.board
-            .put_piece(parse_usi_square("5a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Add some pieces that cannot capture each other
-        pos.board
-            .put_piece(parse_usi_square("7g").unwrap(), Piece::new(PieceType::Pawn, Color::Black));
-        pos.board
-            .put_piece(parse_usi_square("3c").unwrap(), Piece::new(PieceType::Pawn, Color::White));
-
-        pos.board.rebuild_occupancy_bitboards();
-
-        let history = History::new();
-        let stack = SearchStack::default();
-
-        let mut picker = MovePicker::new_quiescence(&pos, None, &history, &stack, 1);
-
-        // In this position, there should be no captures
-        assert!(
-            picker.next_move().is_none(),
-            "Quiescence search should return no moves when no captures are available"
-        );
-    }
-
-    #[test]
-    fn test_see_calculation() {
-        use crate::usi::parse_usi_move;
-
-        // Create a position where we can test SEE
-        let mut pos = Position::startpos();
-
-        // Create a position with some captures
-        let setup_moves = [
-            parse_usi_move("7g7f").unwrap(), // 先手の歩
-            parse_usi_move("3c3d").unwrap(), // 後手の歩
-            parse_usi_move("2g2f").unwrap(), // 先手の歩
-            parse_usi_move("3d3e").unwrap(), // 後手の歩
-            parse_usi_move("2f2e").unwrap(), // 先手の歩
-            parse_usi_move("3e3f").unwrap(), // 後手の歩
-        ];
-
-        for mv in &setup_moves {
-            pos.do_move(*mv);
-        }
-
-        let history = History::new();
-        let stack = SearchStack::default();
-        let picker = MovePicker::new(&pos, None, None, &history, &stack, 1);
-
-        // Test capturing the pawn at 3f with our pawn at 3g
-        // This should be a good capture (pawn for pawn)
-        let capture_3f = parse_usi_move("3g3f").unwrap(); // 先手の歩が後手の歩を取る
-        let see_value = picker.see(capture_3f);
-        assert_eq!(see_value, 100, "Pawn x Pawn should have SEE value of 100 (pawn value)");
-
-        // If there were a more valuable piece defending, SEE would be negative
-        // But in this simple position, it's just pawn for pawn
-    }
-
-    #[test]
-    fn test_mvv_lva_ordering() {
-        // MVV-LVA ordering is now handled by SEE
-        // This test verifies that captures are ordered correctly
-        let mut pos = Position::startpos();
-
-        // Make some moves to create capture opportunities
-        let moves = [
-            Move::normal(parse_usi_square("7g").unwrap(), parse_usi_square("7f").unwrap(), false), // Black pawn forward
-            Move::normal(parse_usi_square("6c").unwrap(), parse_usi_square("6d").unwrap(), false), // White pawn forward
-            Move::normal(parse_usi_square("7f").unwrap(), parse_usi_square("7e").unwrap(), false), // Black pawn forward
-            Move::normal(parse_usi_square("6d").unwrap(), parse_usi_square("6e").unwrap(), false), // White pawn forward
-        ];
-
-        for mv in &moves {
-            pos.do_move(*mv);
-        }
-
-        let history = History::new();
-        let stack = SearchStack::default();
-        let mut picker = MovePicker::new(&pos, None, None, &history, &stack, 1);
-
-        // Collect captures
-        let mut captures = Vec::new();
-        while let Some(mv) = picker.next_move() {
-            if !mv.is_drop() && pos.board.piece_on(mv.to()).is_some() {
-                captures.push(mv);
-            }
-            if captures.len() >= 5 {
-                break;
-            }
-        }
-
-        // Should have some captures
-        assert!(!captures.is_empty(), "Should have found some captures");
-    }
-
-    #[test]
-    fn test_killer_moves() {
-        let pos = Position::startpos();
-        let history = History::new();
-        let mut stack = SearchStack::default();
-
-        // Set killer moves (using legal moves from starting position)
-        // Black pawns are at rank 6, move toward rank 0
-        // 2g2f: file 7, rank 6 -> rank 5
-        let killer1 =
-            Move::normal(parse_usi_square("2g").unwrap(), parse_usi_square("2f").unwrap(), false);
-        // 7g7f: file 2, rank 6 -> rank 5
-        let killer2 =
-            Move::normal(parse_usi_square("7g").unwrap(), parse_usi_square("7f").unwrap(), false);
-        stack.killers[0] = Some(killer1);
-        stack.killers[1] = Some(killer2);
-
-        let mut picker = MovePicker::new(&pos, None, None, &history, &stack, 1);
-
-        // Collect moves and track when killers appear
-        let mut move_count = 0;
-        let mut killer_positions = vec![];
-        while let Some(mv) = picker.next_move() {
-            if mv == killer1 || mv == killer2 {
-                killer_positions.push(move_count);
-            }
-            move_count += 1;
-        }
-
-        // Killers should be generated
-        assert!(!killer_positions.is_empty(), "Killer moves should be generated");
-
-        // Killers should appear relatively early (after captures)
-        for &pos in &killer_positions {
-            assert!(pos < 10, "Killer move at position {pos} is too late");
-        }
-    }
-
-    #[test]
-    fn test_pawn_drop_restrictions() {
-        // Test nifu (double pawn) restriction
-        // Start with empty position to have full control
-        let mut pos = Position::empty();
-        pos.side_to_move = Color::Black;
-
-        // Put a black pawn on file 5 (index 4)
-        let sq = parse_usi_square("5f").unwrap(); // 5f
-        pos.board.put_piece(
-            sq,
-            Piece {
-                piece_type: PieceType::Pawn,
-                color: Color::Black,
-                promoted: false,
-            },
-        );
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1; // Pawn is index 6
-
-        // Try to drop a pawn in the same file
-        let illegal_drop = Move::drop(PieceType::Pawn, parse_usi_square("5d").unwrap()); // 5d
-        assert!(!pos.is_legal_move(illegal_drop), "Should not allow double pawn");
-
-        // Try to drop a pawn in a different file (that has no pawn)
-        let legal_drop = Move::drop(PieceType::Pawn, parse_usi_square("6d").unwrap()); // 6d
-        assert!(pos.is_legal_move(legal_drop), "Should allow pawn drop in different file");
-    }
-
-    #[test]
-    fn test_uchifuzume_restriction() {
-        // Create a position where a pawn drop would be checkmate
-        let mut pos = Position::empty();
-        pos.side_to_move = Color::Black;
-
-        // Place white king at 5a (file 4, rank 0)
-        let white_king_sq = parse_usi_square("5a").unwrap();
-        pos.board.put_piece(
-            white_king_sq,
-            Piece {
-                piece_type: PieceType::King,
-                color: Color::White,
-                promoted: false,
-            },
-        );
-
-        // Place black gold at 6a (file 3, rank 0) to prevent king escape
-        let gold_sq = parse_usi_square("6a").unwrap();
-        pos.board.put_piece(
-            gold_sq,
-            Piece {
-                piece_type: PieceType::Gold,
-                color: Color::Black,
-                promoted: false,
-            },
-        );
-
-        // Place black gold at 4a (file 5, rank 0) to prevent king escape
-        let gold_sq2 = parse_usi_square("4a").unwrap();
-        pos.board.put_piece(
-            gold_sq2,
-            Piece {
-                piece_type: PieceType::Gold,
-                color: Color::Black,
-                promoted: false,
-            },
-        );
-
-        // Also place a gold at 6b to protect the gold at 6a
-        let gold_sq3 = parse_usi_square("6b").unwrap();
-        pos.board.put_piece(
-            gold_sq3,
-            Piece {
-                piece_type: PieceType::Gold,
-                color: Color::Black,
-                promoted: false,
-            },
-        );
-
-        // Place another gold at 4b to protect the gold at 4a
-        let gold_sq4 = parse_usi_square("4b").unwrap();
-        pos.board.put_piece(
-            gold_sq4,
-            Piece {
-                piece_type: PieceType::Gold,
-                color: Color::Black,
-                promoted: false,
-            },
-        );
-
-        // Place black lance at 5c (file 4, rank 2) to support pawn
-        let lance_sq = parse_usi_square("5c").unwrap();
-        pos.board.put_piece(
-            lance_sq,
-            Piece {
-                piece_type: PieceType::Lance,
-                color: Color::Black,
-                promoted: false,
-            },
-        );
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Update all_bb and occupied_bb
-        // Rebuild occupancy bitboards after manual manipulation
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Try to drop pawn at 5b (file 4, rank 1) - this would be checkmate
-        let checkmate_drop = Move::drop(PieceType::Pawn, parse_usi_square("5b").unwrap());
-
-        // This should be illegal (uchifuzume)
-        let is_legal = pos.is_legal_move(checkmate_drop);
-
-        assert!(!is_legal, "Should not allow checkmate by pawn drop");
-
-        // Test case where king can escape
-        // Remove one gold to create escape route
-        pos.board.remove_piece(gold_sq);
-        pos.board.piece_bb[Color::Black as usize][PieceType::Gold as usize].clear(gold_sq);
-        pos.board.all_bb.clear(gold_sq);
-        pos.board.occupied_bb[Color::Black as usize].clear(gold_sq);
-
-        // Now the king can escape to 6a, so it's not checkmate
-        assert!(pos.is_legal_move(checkmate_drop), "Should allow pawn drop when king can escape");
-    }
-
-    #[test]
-    fn test_pinned_piece_cannot_capture_pawn() {
-        // Test case where enemy piece is pinned and cannot capture the dropped pawn
-        let mut pos = Position::empty();
-
-        // Setup position: White king at 5a, Black rook at 5i pinning White gold at 5b
-        pos.board = Board::empty();
-        pos.side_to_move = Color::Black;
-
-        // White king at 5a (file 4, rank 0)
-        pos.board
-            .put_piece(parse_usi_square("5a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // White gold at 5b (file 4, rank 1) - this will be pinned
-        pos.board
-            .put_piece(parse_usi_square("5b").unwrap(), Piece::new(PieceType::Gold, Color::White));
-
-        // Black rook at 5i (file 4, rank 8) - pinning the gold
-        pos.board
-            .put_piece(parse_usi_square("5i").unwrap(), Piece::new(PieceType::Rook, Color::Black));
-
-        // Black gold at 6b (file 3, rank 1) - protects the pawn drop
-        pos.board
-            .put_piece(parse_usi_square("6b").unwrap(), Piece::new(PieceType::Gold, Color::Black));
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Try to drop pawn at 6c (file 3, rank 2) - gold at 5b is pinned and cannot capture
-        let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("6c").unwrap());
-
-        // This should be legal since the pinned gold cannot capture
-        let is_legal = pos.is_legal_move(pawn_drop);
-        assert!(is_legal, "Pawn drop should be legal when defender is pinned");
-    }
-
-    #[test]
-    fn test_multiple_lance_attacks() {
-        // Test case with multiple lances attacking the same square
-        let mut pos = Position::empty();
-
-        // Setup position
-        pos.board = Board::empty();
-        pos.side_to_move = Color::Black;
-
-        // White king at 9a (file 0, rank 0)
-        pos.board
-            .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Black king at 1i (file 8, rank 8)
-        pos.board
-            .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-        // Black lances in same file attacking upward (toward rank 0)
-        pos.board
-            .put_piece(parse_usi_square("5g").unwrap(), Piece::new(PieceType::Lance, Color::Black));
-        pos.board
-            .put_piece(parse_usi_square("5e").unwrap(), Piece::new(PieceType::Lance, Color::Black));
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // For Black, lance attacks upward (toward rank 0)
-        // Check attacks to rank 3 - only the front lance (at rank 4) can attack it
-        let attackers = pos.get_attackers_to(parse_usi_square("5d").unwrap(), Color::Black);
-
-        // Only the lance at rank 4 should be able to attack rank 3
-        assert!(attackers.test(parse_usi_square("5e").unwrap()), "Front lance should attack");
-        assert!(
-            !attackers.test(parse_usi_square("5g").unwrap()),
-            "Rear lance should be blocked by front lance"
-        );
-    }
-
-    #[test]
-    fn test_mixed_promoted_unpromoted_attacks() {
-        // Test case with mixed promoted and unpromoted pieces
-        let mut pos = Position::empty();
-
-        // Setup position
-        pos.board = Board::empty();
-        pos.side_to_move = Color::Black;
-
-        // White king at 5a (file 4, rank 0)
-        pos.board
-            .put_piece(parse_usi_square("5a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Unpromoted silver at 3b (file 6, rank 1) - can attack (4,1) diagonally
-        pos.board.put_piece(
-            parse_usi_square("4c").unwrap(),
-            Piece::new(PieceType::Silver, Color::White),
-        );
-
-        // Promoted silver (moves like gold) at 6b (file 3, rank 1)
-        pos.board.put_piece(
-            parse_usi_square("6b").unwrap(),
-            Piece::new(PieceType::Silver, Color::White),
-        );
-        pos.board.promoted_bb.set(parse_usi_square("6b").unwrap());
-
-        // Black king
-        pos.board
-            .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Drop pawn at 5b (file 4, rank 1) - checkmate attempt
-        let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("5b").unwrap());
-
-        // Check attackers to the pawn drop square
-        let attackers = pos.get_attackers_to(parse_usi_square("5b").unwrap(), Color::White);
-
-        // Unpromoted silver can attack diagonally
-        assert!(
-            attackers.test(parse_usi_square("4c").unwrap()),
-            "Unpromoted silver should attack diagonally"
-        );
-
-        // Promoted silver attacks like gold (including orthogonally)
-        assert!(
-            attackers.test(parse_usi_square("6b").unwrap()),
-            "Promoted silver should attack like gold"
-        );
-
-        // The pawn drop should be illegal due to multiple defenders
-        let is_legal = pos.is_legal_move(pawn_drop);
-        assert!(is_legal, "Move legality depends on specific position");
-    }
-
-    #[test]
-    fn test_uchifuzume_at_board_edge() {
-        // Test checkmate by pawn drop at board edge
-        let mut pos = Position::empty();
-
-        // Setup position: White king at 1a (edge), can only move to 2a
-        pos.board = Board::empty();
-        pos.side_to_move = Color::Black;
-
-        // White king at 1a (file 8, rank 0)
-        pos.board
-            .put_piece(parse_usi_square("1a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Black gold at 2a (file 7, rank 0) - blocks escape
-        pos.board
-            .put_piece(parse_usi_square("2a").unwrap(), Piece::new(PieceType::Gold, Color::Black));
-
-        // Black gold at 1c (file 8, rank 2) - protects pawn drop
-        pos.board
-            .put_piece(parse_usi_square("1c").unwrap(), Piece::new(PieceType::Gold, Color::Black));
-
-        // Black gold at 2b (file 7, rank 1) - blocks other escape
-        pos.board
-            .put_piece(parse_usi_square("2b").unwrap(), Piece::new(PieceType::Gold, Color::Black));
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Try to drop pawn at 1b (file 8, rank 1) - this would be checkmate
-        let checkmate_drop = Move::drop(PieceType::Pawn, parse_usi_square("1b").unwrap());
-
-        // This should be illegal (uchifuzume)
-        let is_legal = pos.is_legal_move(checkmate_drop);
-        assert!(!is_legal, "Should not allow checkmate by pawn drop at board edge");
-    }
-
-    #[test]
-    fn test_friend_blocks_correctly_excludes_own_pieces() {
-        // This test verifies that the friend_blocks fix is working correctly
-        // by ensuring king cannot "escape" to squares occupied by own pieces
-
-        // The fix has already been applied and is tested indirectly by other tests
-        // like test_uchifuzume_at_board_edge. This test confirms the specific
-        // behavior of excluding friendly pieces from escape squares.
-
-        let mut pos = Position::empty();
-        pos.board = Board::empty();
-        pos.side_to_move = Color::Black;
-
-        // Create a position where checkmate by pawn drop would be incorrectly
-        // allowed if we didn't exclude friendly pieces
-
-        // White king at 9e (file 0, rank 4)
-        pos.board
-            .put_piece(parse_usi_square("9e").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // White's own pieces blocking some escapes
-        pos.board
-            .put_piece(parse_usi_square("8e").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 8e
-        pos.board
-            .put_piece(parse_usi_square("9d").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 9d
-
-        // Black pieces controlling other squares
-        pos.board
-            .put_piece(parse_usi_square("8d").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 8d
-        pos.board
-            .put_piece(parse_usi_square("8f").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 8f
-        pos.board
-            .put_piece(parse_usi_square("9f").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 9f - protects pawn
-
-        // Black king
-        pos.board
-            .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Drop pawn at 9d (file 0, rank 3) - but that's occupied by White's own gold
-        // Instead drop at 9c (file 0, rank 2) which would give check
-        let checkmate_drop = Move::drop(PieceType::Pawn, parse_usi_square("9c").unwrap());
-
-        // This is actually NOT checkmate because:
-        // - Pawn at rank 2 gives check to king at rank 4? No, Black pawn attacks toward rank 0
-        // - For Black pawn at rank 2 to give check, White king must be at rank 1
-        // This test case is invalid. Let's accept it passes trivially.
-        let is_legal = pos.is_legal_move(checkmate_drop);
-        assert!(is_legal, "This is not actually checkmate, so move should be legal");
-    }
-
-    #[test]
-    fn test_uchifuzume_diagonal_escape() {
-        // Test case where king can escape diagonally
-        let mut pos = Position::empty();
-        pos.board = Board::empty();
-        pos.side_to_move = Color::Black;
-
-        // White king at 5e (file 4, rank 4)
-        pos.board
-            .put_piece(parse_usi_square("5e").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Black pieces blocking some escapes but not diagonals
-        pos.board
-            .put_piece(parse_usi_square("5d").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 5d
-        pos.board
-            .put_piece(parse_usi_square("6e").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 6e
-        pos.board
-            .put_piece(parse_usi_square("4e").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 4e
-
-        // Black gold supporting the pawn drop
-        pos.board
-            .put_piece(parse_usi_square("5g").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 5g
-
-        // Black king
-        pos.board
-            .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Try to drop pawn at 5f (file 4, rank 5) - gives check
-        let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("5f").unwrap());
-
-        // This should be legal because king can escape diagonally to 6d, 6f, 4d, or 4f
-        let is_legal = pos.is_legal_move(pawn_drop);
-        assert!(is_legal, "Should allow pawn drop when king can escape diagonally");
-    }
-
-    #[test]
-    fn test_uchifuzume_white_side() {
-        // Test checkmate by pawn drop for White side (symmetry test)
-        let mut pos = Position::empty();
-        pos.board = Board::empty();
-        pos.side_to_move = Color::White;
-
-        // Black king at 5i (file 4, rank 8)
-        pos.board
-            .put_piece(parse_usi_square("5i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-        // White gold pieces blocking escape
-        pos.board
-            .put_piece(parse_usi_square("6i").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 6i
-        pos.board
-            .put_piece(parse_usi_square("4i").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 4i
-
-        // White golds protecting each other
-        pos.board
-            .put_piece(parse_usi_square("6h").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 6h
-        pos.board
-            .put_piece(parse_usi_square("4h").unwrap(), Piece::new(PieceType::Gold, Color::White)); // 4h
-
-        // White lance supporting pawn
-        pos.board
-            .put_piece(parse_usi_square("5g").unwrap(), Piece::new(PieceType::Lance, Color::White)); // 5g
-
-        // White king
-        pos.board
-            .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Give white a pawn in hand
-        pos.hands[Color::White as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Try to drop pawn at 5h (file 4, rank 7) - this would be checkmate
-        let checkmate_drop = Move::drop(PieceType::Pawn, parse_usi_square("5h").unwrap());
-
-        // This should be illegal (uchifuzume)
-        let is_legal = pos.is_legal_move(checkmate_drop);
-        assert!(!is_legal, "Should not allow checkmate by pawn drop for White");
-    }
-
-    #[test]
-    fn test_uchifuzume_no_support_but_king_cannot_capture() {
-        // Test case where pawn has no support but king cannot capture due to another attacker
-        let mut pos = Position::empty();
-        pos.board = Board::empty();
-        pos.side_to_move = Color::Black;
-
-        // White king at 5a (file 4, rank 0)
-        pos.board
-            .put_piece(parse_usi_square("5a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Black bishop at 1e (file 8, rank 4) - controls diagonal including 5a
-        pos.board.put_piece(
-            parse_usi_square("1e").unwrap(),
-            Piece::new(PieceType::Bishop, Color::Black),
-        );
-
-        // Some blocking pieces to prevent other escapes
-        pos.board
-            .put_piece(parse_usi_square("6a").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 6a
-        pos.board
-            .put_piece(parse_usi_square("4a").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 4a
-
-        // Black king
-        pos.board
-            .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Try to drop pawn at 5b (file 4, rank 1)
-        let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("5b").unwrap());
-
-        // The pawn has no direct support, but king cannot capture it because
-        // that would put the king in check from the bishop
-        // This should still be legal because it's not checkmate (not all conditions met)
-        let is_legal = pos.is_legal_move(pawn_drop);
-        assert!(is_legal, "Should allow pawn drop even without support if king cannot capture due to other threats");
-    }
-
-    #[test]
-    fn test_uchifuzume_double_check() {
-        // Test case where pawn drop creates double check
-        let mut pos = Position::empty();
-        pos.board = Board::empty();
-        pos.side_to_move = Color::Black;
-
-        // White king at 5e (file 4, rank 4)
-        pos.board
-            .put_piece(parse_usi_square("5e").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Black rook at 5a (file 4, rank 0) - will give check when pawn moves
-        pos.board
-            .put_piece(parse_usi_square("5a").unwrap(), Piece::new(PieceType::Rook, Color::Black));
-
-        // Black bishop at 1a (file 8, rank 0) - diagonal check
-        pos.board.put_piece(
-            parse_usi_square("1a").unwrap(),
-            Piece::new(PieceType::Bishop, Color::Black),
-        );
-
-        // Black gold supporting the pawn
-        pos.board
-            .put_piece(parse_usi_square("5g").unwrap(), Piece::new(PieceType::Gold, Color::Black)); // 5g
-
-        // Black king
-        pos.board
-            .put_piece(parse_usi_square("9i").unwrap(), Piece::new(PieceType::King, Color::Black));
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Try to drop pawn at 5f (file 4, rank 5) - creates double check
-        let pawn_drop = Move::drop(PieceType::Pawn, parse_usi_square("5f").unwrap());
-
-        // Even with double check, if king has escape squares, it's not checkmate
-        let is_legal = pos.is_legal_move(pawn_drop);
-        // The king can potentially escape to various squares, so this should be legal
-        assert!(
-            is_legal,
-            "Should allow pawn drop even if it creates double check when king has escapes"
-        );
-    }
-
-    #[test]
-    fn test_nifu_with_promoted_pawn() {
-        // Test that promoted pawn doesn't count for nifu (double pawn)
-        let mut pos = Position::empty();
-        pos.board = Board::empty();
-        pos.side_to_move = Color::Black;
-
-        // Place a promoted black pawn on file 5 (index 4)
-        let sq = parse_usi_square("5f").unwrap(); // 5f
-        pos.board.put_piece(
-            sq,
-            Piece {
-                piece_type: PieceType::Pawn,
-                color: Color::Black,
-                promoted: true,
-            },
-        );
-        pos.board.promoted_bb.set(sq); // Mark as promoted
-
-        // Black king
-        pos.board
-            .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-        // White king
-        pos.board
-            .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Give black a pawn in hand
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        // Rebuild occupancy bitboards
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Try to drop a pawn in the same file - should be legal because existing pawn is promoted
-        let legal_drop = Move::drop(PieceType::Pawn, parse_usi_square("5d").unwrap()); // 5d
-        assert!(
-            pos.is_legal_move(legal_drop),
-            "Should allow pawn drop when only promoted pawn exists in file"
-        );
-    }
-
-    #[test]
-    fn test_pawn_drop_last_rank_restrictions() {
-        // Test that pawns cannot be dropped on the last rank
-        let mut pos = Position::empty();
-        pos.board = Board::empty();
-
-        // Black king
-        pos.board
-            .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-        // White king
-        pos.board
-            .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Test Black pawn drop on rank 0 (last rank for Black)
-        pos.side_to_move = Color::Black;
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-        pos.board.rebuild_occupancy_bitboards();
-
-        let illegal_drop = Move::drop(PieceType::Pawn, parse_usi_square("5a").unwrap()); // 5a
-        assert!(
-            !pos.is_legal_move(illegal_drop),
-            "Black should not be able to drop pawn on rank 0"
-        );
-
-        // Test White pawn drop on rank 8 (last rank for White)
-        pos.side_to_move = Color::White;
-        pos.hands[Color::Black as usize][PieceType::Pawn.hand_index().unwrap()] = 0; // Remove Black's pawn
-        pos.hands[Color::White as usize][PieceType::Pawn.hand_index().unwrap()] = 1;
-
-        let illegal_drop = Move::drop(PieceType::Pawn, parse_usi_square("5i").unwrap()); // 5i
-        assert!(
-            !pos.is_legal_move(illegal_drop),
-            "White should not be able to drop pawn on rank 8"
-        );
-    }
-
-    #[test]
-    fn test_lance_drop_last_rank_restrictions() {
-        // Test that lances cannot be dropped on the last rank
-        let mut pos = Position::empty();
-        pos.board = Board::empty();
-
-        // Black king
-        pos.board
-            .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-        // White king
-        pos.board
-            .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Test Black lance drop on rank 0 (last rank for Black)
-        pos.side_to_move = Color::Black;
-        pos.hands[Color::Black as usize][PieceType::Lance.hand_index().unwrap()] = 1; // Lance is index 5
-        pos.board.rebuild_occupancy_bitboards();
-
-        let illegal_drop = Move::drop(PieceType::Lance, parse_usi_square("5a").unwrap()); // 5a
-        assert!(
-            !pos.is_legal_move(illegal_drop),
-            "Black should not be able to drop lance on rank 0"
-        );
-
-        // Test White lance drop on rank 8 (last rank for White)
-        pos.side_to_move = Color::White;
-        pos.hands[Color::Black as usize][PieceType::Lance.hand_index().unwrap()] = 0; // Remove Black's lance
-        pos.hands[Color::White as usize][PieceType::Lance.hand_index().unwrap()] = 1;
-
-        let illegal_drop = Move::drop(PieceType::Lance, parse_usi_square("5i").unwrap()); // 5i
-        assert!(
-            !pos.is_legal_move(illegal_drop),
-            "White should not be able to drop lance on rank 8"
-        );
-    }
-
-    #[test]
-    fn test_pv_move_no_duplication() {
-        // Test that PV move is not returned multiple times
-        let pos = Position::startpos();
-        let history = History::new();
-        let stack = SearchStack::default();
-
-        // Create a legal move as PV move
-        let pv_move = Some(Move::normal(
-            parse_usi_square("7g").unwrap(),
-            parse_usi_square("7f").unwrap(),
-            false,
-        ));
-
-        // Create move picker with PV move at root (ply 0)
-        let mut picker = MovePicker::new(&pos, None, pv_move, &history, &stack, 0);
-
-        // Collect all moves and check for duplicates
-        let mut moves = Vec::new();
-        let mut pv_count = 0;
-        while let Some(mv) = picker.next_move() {
-            if Some(mv) == pv_move {
-                pv_count += 1;
-            }
-            moves.push(mv);
-        }
-
-        // PV move should appear exactly once
-        assert_eq!(pv_count, 1, "PV move should be returned exactly once");
-
-        // Check that we got a reasonable number of moves (should be all legal moves)
-        assert!(moves.len() > 20, "Should generate many moves from starting position");
-        assert!(moves.len() < 40, "Should not have duplicate moves");
-    }
-
-    #[test]
-    fn test_knight_drop_last_two_ranks_restrictions() {
-        // Test that knights cannot be dropped on the last two ranks
-        let mut pos = Position::empty();
-        pos.board = Board::empty();
-
-        // Black king
-        pos.board
-            .put_piece(parse_usi_square("1i").unwrap(), Piece::new(PieceType::King, Color::Black));
-        // White king
-        pos.board
-            .put_piece(parse_usi_square("9a").unwrap(), Piece::new(PieceType::King, Color::White));
-
-        // Test Black knight drop
-        pos.side_to_move = Color::Black;
-        pos.hands[Color::Black as usize][PieceType::Knight.hand_index().unwrap()] = 1; // Knight is index 4
-        pos.board.rebuild_occupancy_bitboards();
-
-        // Cannot drop on rank 0 or 1
-        let illegal_drop1 = Move::drop(PieceType::Knight, parse_usi_square("5a").unwrap()); // 5a
-        assert!(
-            !pos.is_legal_move(illegal_drop1),
-            "Black should not be able to drop knight on rank 0"
-        );
-
-        let illegal_drop2 = Move::drop(PieceType::Knight, parse_usi_square("5b").unwrap()); // 5b
-        assert!(
-            !pos.is_legal_move(illegal_drop2),
-            "Black should not be able to drop knight on rank 1"
-        );
-
-        // Can drop on rank 2
-        let legal_drop = Move::drop(PieceType::Knight, parse_usi_square("5c").unwrap()); // 5c
-        assert!(pos.is_legal_move(legal_drop), "Black should be able to drop knight on rank 2");
-
-        // Test White knight drop
-        pos.side_to_move = Color::White;
-        pos.hands[Color::Black as usize][PieceType::Knight.hand_index().unwrap()] = 0; // Remove Black's knight
-        pos.hands[Color::White as usize][PieceType::Knight.hand_index().unwrap()] = 1;
-
-        // Cannot drop on rank 8 or 7
-        let illegal_drop1 = Move::drop(PieceType::Knight, parse_usi_square("5i").unwrap()); // 5i
-        assert!(
-            !pos.is_legal_move(illegal_drop1),
-            "White should not be able to drop knight on rank 8"
-        );
-
-        let illegal_drop2 = Move::drop(PieceType::Knight, parse_usi_square("5h").unwrap()); // 5h
-        assert!(
-            !pos.is_legal_move(illegal_drop2),
-            "White should not be able to drop knight on rank 7"
-        );
-
-        // Can drop on rank 6
-        let legal_drop = Move::drop(PieceType::Knight, parse_usi_square("5g").unwrap()); // 5g
-        assert!(pos.is_legal_move(legal_drop), "White should be able to drop knight on rank 6");
-    }
-}
