+++ packages/rust-core/crates/engine-core/src/movegen/move_picker/mod.rs
@@ -1,0 +1,16 @@
+//! Move picker module for efficient move ordering
+
+mod move_generation;
+mod move_scoring;
+mod picker;
+mod position_ext;
+mod types;
+
+// Re-export public types
+pub use picker::MovePicker;
+// Internal types are available through super imports
+
+// Position extensions are available internally
+
+#[cfg(test)]
+mod tests;
