+++ packages/rust-core/crates/engine-core/src/movegen/move_picker/tests/mod.rs
@@ -1,0 +1,6 @@
+//! Tests for move picker module
+
+mod basic_tests;
+mod capture_tests;
+mod drop_tests;
+mod uchifuzume_tests;
